+++ univariate_opt/src/univariate_opt.h
@@ -17,7 +17,7 @@ double numerical_hessian(double (*opt_fun)(double), double x, double eps);
 struct OptResults gradient_descent_1d(double (*opt_fun)(double), double x_guess, double gamma, int max_iter, double tol);
 struct OptResults newton_1d(double (*opt_fun)(double), double x_guess, int max_iter, double tol);
 struct OptResults quadratic_fit_search(double (*opt_fun)(double), double a, double b, double c, int max_inter, double tol);
-float evaluate_sign(double x);
+int evaluate_sign(double x);
 struct OptResults bissection_method(double (*opt_fun)(double), double lower_lim, double upper_lim, int max_inter, double tol);
 
 #endif
